requirements.txt
discord.py>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
pymongo>=4.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.2.0
python-multipart>=0.0.5
motor>=3.3.0
wavelink>=3.0.0
pynacl>=1.5.0
